refactor(browser): neutralize browser computer-use and host route shell naming
Rename the five internal API route shells and their route-array exports while preserving every runtime and protocol identity.
